API

Modules

Types and constants

Functions and macros

MJDSigGen.nearest_field_grid_index

Documentation

MJDSigGen.nearest_field_grid_index — Method

nearest_field_grid_index(setup::SigGenSetup, location::NTuple{3, Real})

Returns:

(:outside, i, j), if outside crystal or too far from a valid grid point

(:interpol, i, j, if interpolation is okay

(:extrapol, i, j), if we can find a point but extrapolation is needed